--- src/hooks/use-check-is-mobile.ts
@@ -1,11 +1,0 @@
-import { useState } from 'react';
-import { isMobileCheck } from '../utils';
-
-/**
- * Returns if the browser is a mobile browser.
- */
-export const useCheckIsMobile = (): { isMobile: boolean } => {
-  const [isMobile] = useState(isMobileCheck());
-
-  return { isMobile };
-};
