--- src/hooks/use-dispatch.ts
@@ -1,8 +1,0 @@
-import { useDispatch as useTypedDispatch } from 'react-redux';
-import { Dispatch } from '../data';
-
-/**
- * Typed useDispatch hook.
- */
-export const useDispatch = (): ReturnType<typeof useTypedDispatch> =>
-  useTypedDispatch<Dispatch>();
