+++ src/utils/index.ts
@@ -2,6 +2,7 @@ export * from './keybinds';
 export * from './message';
 export * from './page';
 export * from './player';
+export * from './redux';
 export * from './responsive';
 export * from './shadow-dom';
 export * from './string';
